Added more intersections in hallways
Previously the maze generation was done using Prim's algorithm, which creates a tree that connects all of the nodes in the graph (the maze acted as a graph). However, this resulted in there only being one path between opponents (with all other branches leading to dead-ends). This changes the maze generation by adding a loop after the initial maze generation that "pokes holes" in the walls to create multiple paths between players, making gameplay more interesting