Fixed issue with maze generation
Previously the "poke hole" algorithm would poke a hole in the wall if one side was cellChar even if the other side was wallChar, resulting in a small chance of having weird indentations in the walls. This fixes that issue by checking that both sides of the "punched hole" are cellChar instead of just one side.